+++ pytorch_lightning/plugins/training_type/ddp_spawn.py
@@ -14,6 +14,7 @@
 import logging
 import os
 import re
+from contextlib import contextmanager
 from multiprocessing.queues import SimpleQueue
 from typing import Any, Dict, List, Optional, Union
 
@@ -364,3 +365,16 @@ class DDPSpawnPlugin(ParallelPlugin):
             description="DDPSpawn Plugin with `find_unused_parameters` as False",
             find_unused_parameters=False,
         )
+
+    @contextmanager
+    def block_backward_sync(self):
+        """
+        Blocks ddp sync gradients behaviour on backwards pass.
+        This is useful for skipping sync when accumulating gradients, reducing communication overhead
+        Returns: context manager with sync behaviour off
+        """
+        if isinstance(self.model, DistributedDataParallel):
+            with self.model.no_sync():
+                yield None
+        else:
+            yield None